// src/project/log/console-types.d.ts
export declare enum ConsoleCode {
    Reset = "\u001B[0m",
    Bright = "\u001B[1m",
    Dim = "\u001B[2m",
    Underscore = "\u001B[4m",
    Blink = "\u001B[5m",
    Reverse = "\u001B[7m",
    Hidden = "\u001B[8m"
}
export declare enum ForegroundColor {
    _8_ForegroundBlack = "\u001B[30m",
    _8_ForegroundRed = "\u001B[31m",
    _8_ForegroundGreen = "\u001B[32m",
    _8_ForegroundYellow = "\u001B[33m",
    _8_ForegroundBlue = "\u001B[34m",
    _8_ForegroundMagenta = "\u001B[35m",
    _8_ForegroundCyan = "\u001B[36m",
    _8_ForegroundWhite = "\u001B[37m",
    _256_ForegroundBlueish27 = "\u001B[38;5;27m",
    _256_ForegroundDullGreen29 = "\u001B[38;5;29m",
    _256_ForegroundGrayish101 = "\u001B[38;5;101m",
    _256_ForegroundOrange208 = "\u001B[38;5;208m"
}
export declare enum BackgroundColor {
    _8_BackgroundBlack = "\u001B[40m",
    _8_BackgroundRed = "\u001B[41m",
    _8_BackgroundGreen = "\u001B[42m",
    _8_BackgroundYellow = "\u001B[43m",
    _8_BackgroundBlue = "\u001B[44m",
    _8_BackgroundMagenta = "\u001B[45m",
    _8_BackgroundCyan = "\u001B[46m",
    _8_BackgroundWhite = "\u001B[47m",
    _256_Blueish27 = "\u001B[48;5;27m",
    _256_BackgroundGrayish101 = "\u001B[48;5;101m",
    _256_BackgroundDullGreen29 = "\u001B[48;5;29m",
    _256_BackgroundOrange208 = "\u001B[48;5;208m"
}
export declare enum utf8SpecialCharacters {
    RighwardsArrow = "\u2192",
    WhiteRightArrow = "\u21E8",
    WhiteHeavyCheckMark = "\u2705",
    HeavyCheckmark = "\u2714"
}